+++ words.txt
@@ -1,4 +1,4 @@
 hello|привет|0
-dog|собака|0
-cat|кошка|0
-thank you|спасибо|0
+dog|собака|3
+cat|кошка|1
+thank you|спасибо|2
